Move layout_weight to ViewProxy so that LinearLayout children can set weight
