+++ paris-test/src/test/resources/default_values/Paris.java
@@ -7,8 +7,6 @@ import android.view.ViewGroupStyleApplier;
 import android.view.ViewStyleApplier;
 import android.widget.ImageView;
 import android.widget.ImageViewStyleApplier;
-import android.widget.LinearLayout;
-import android.widget.LinearLayoutStyleApplier;
 import android.widget.TextView;
 import android.widget.TextViewStyleApplier;
 import com.airbnb.paris.spannables.SpannableBuilder;
@@ -20,14 +18,6 @@ public final class Paris {
 
   public static ImageViewStyleApplier.StyleBuilder styleBuilder(ImageView view) {
     return new ImageViewStyleApplier.StyleBuilder(new ImageViewStyleApplier(view));
-  }
-
-  public static LinearLayoutStyleApplier style(LinearLayout view) {
-    return new LinearLayoutStyleApplier(view);
-  }
-
-  public static LinearLayoutStyleApplier.StyleBuilder styleBuilder(LinearLayout view) {
-    return new LinearLayoutStyleApplier.StyleBuilder(new LinearLayoutStyleApplier(view));
   }
 
   public static MyOtherViewStyleApplier style(MyOtherView view) {
@@ -79,7 +69,6 @@ public final class Paris {
    */
   public static void assertStylesContainSameAttributes(Context context) {
     ImageViewStyleApplier.assertStylesContainSameAttributes(context);
-    LinearLayoutStyleApplier.assertStylesContainSameAttributes(context);
     MyOtherViewStyleApplier.assertStylesContainSameAttributes(context);
     MyViewStyleApplier.assertStylesContainSameAttributes(context);
     TextViewStyleApplier.assertStylesContainSameAttributes(context);
